==> 19.cibersort_immuneCor.R <==


#if (!requireNamespace("BiocManager", quietly = TRUE))
#    install.packages("BiocManager")
#BiocManager::install("limma")

#install.packages("reshape2")
#install.packages("ggpubr")
#install.packages("ggExtra")

library(limma)
library(reshape2)
library(ggpubr)
library(ggExtra)
##"PIRT" should be replaced during each turn of analysis
gene="PIRT" 
expFile="normalize.txt"    
immFile="CIBERSORT-Results.txt" 
setwd=("C:\\Users\\anxin\\Desktop\\integrated_model") 

rt=read.table("C:\\Users\\anxin\\Desktop\\integrated_model\\normalize.txt", header=T, sep="\t", check.names=F)
rt <- rt[!is.na(rt[, 1]), ]
rt=as.matrix(rt)
rownames(rt)=rt[,1]
exp=rt[,2:ncol(rt)]
dimnames=list(rownames(exp),colnames(exp))
data=matrix(as.numeric(as.matrix(exp)),nrow=nrow(exp),dimnames=dimnames)
data=avereps(data)


data=t(data[gene,,drop=F])
data=as.data.frame(data)


immune=read.table("C:\\Users\\anxin\\Desktop\\integrated_model\\CIBERSORT-Results.txt", header=T, sep="\t", check.names=F, row.names=1)


immune <- immune[complete.cases(immune), ]


immune <- immune[, apply(immune, 2, sd, na.rm=TRUE) != 0]
samele=intersect(row.names(immune), row.names(data))
rt=cbind(immune[samele,], data[samele,])
colnames(rt)[ncol(rt)] <- "PIRT" 
#Create a folder named PIRT in the current directory.
setwd=("C:\\Users\\anxin\\Desktop\\integrated_model\\cor_PIRT")
outTab=data.frame()
for(i in colnames(rt)[1:(ncol(rt)-1)]){
	x=as.numeric(rt[,gene])
	y=as.numeric(rt[,i])
	if(sd(y)==0){y[1]=0.00001}
	cor=cor.test(x, y, method="spearman")
	
	outVector=cbind(Gene=gene, Cell=i, cor=cor$estimate, pvalue=cor$p.value)
	outTab=rbind(outTab,outVector)
	
	if(cor$p.value<0.05){
		outFile=paste0("cor.", i, ".pdf")
		df1=as.data.frame(cbind(x,y))
		p1=ggplot(df1, aes(x, y)) + 
				  xlab(paste0(gene, " expression")) + ylab(i)+
				  geom_point() + geom_smooth(method="lm",formula = y ~ x) + theme_bw()+
				  stat_cor(method = 'spearman', aes(x =x, y =y))
		p2=ggMarginal(p1, type="density", xparams=list(fill = "orange"), yparams=list(fill = "blue"))

		pdf(file=outFile, width=5.2, height=5)
		print(p2)
		dev.off()
	}
}

write.table(outTab,file="C:\\Users\\anxin\\Desktop\\integrated_model\\cor.PIRTult.txt",sep="\t",row.names=F,quote=F)
